backend/requirements.txt
Django==2.2.6
django-extensions==2.2.5
django-environ==0.4.5
djangorestframework==3.10.3
psycopg2==2.8.4
django-enumfield==2.0.1
ipdb

